scanner/requirements.txt
# Core Flask dependencies
Flask==3.0.2
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3

# HTTP request dependencies
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
urllib3==2.3.0

# ZAP dependencies
python-owasp-zap-v2.4==0.1.0

# CORS dependency
flask-cors==5.0.1
# Optional: For production deployment
# gunicorn==22.0.0